FastApi/requirements.txt
sqlalchemy>=2.0.20,<2.1.0
psycopg2-binary
pydantic
python-dotenv
python-multipart
fastapi
uvicorn[standard]
